changing blog-cards to cards, fixing PR comments
