--- blocks/blog-cards/blog-cards.css
@@ -1,175 +1,0 @@
-main a {
-    color: #4a4a4a;    
-}
-
-@media (min-width: 400px) {
-.blog-cards > ul {
-    list-style: none;
-    margin: 0;                    
-    font-size: 1em;
-    font-weight: 400;
-    line-height: 1.5;
-    display: block;
-    padding-inline-start: 0;
-    min-height: 100vh;
-  }
-  
-  .blog-cards > ul > li {     
-    border: 1px solid #c4d1e9;
-    border-radius: 6px;
-    padding: 0;
-    margin-bottom: 1.5rem;    
-  }
-  
-  .blog-cards .blogs-card-body {
-    padding: 1.5rem;
-    display: flex;    
-    flex-direction: column;    
-    justify-content: space-between;    
-    background-color: transparent;
-    min-height: 220px;  
-  } 
-  
-  .blog-cards .blogs-card-image {
-    line-height: 0;
-    position: relative;
-    display: block;
-   
-    
-  }
-   
-  .blog-cards > ul > li img {
-    border-top-left-radius: 6px;
-    border-top-right-radius: 6px;
-    aspect-ratio: 4 / 3;
-    max-width: 100%;
-    object-fit: cover
-  }
-  
-  .blog-cards .blogs-card-body .h3{
-    font-size: var(--body-font-size-l);
-    margin-bottom: 10px;    
-    font-family: var(--heading-font-family);
-    font-weight: 800;
-    color: #28061e;
-  }
-
-  .blog-cards .blogs-card-body >span{
-    font-size: var(--body-font-size-s);
-  }
-
-  .blog-cards .blogs-card-category{
-    font-family: Raleway,sans-serif;
-    font-weight: 800;
-    position: absolute;
-    line-height: 1.7;
-    top: 10px;
-    left: 20px;
-    border-radius: 14px;
-    display: inline-block;
-    padding: 0 14px;    
-    color: #fff;
-    max-width: calc(100% - 2rem);
-  }
-
-  .blog-cards .pagination{
-    margin-top: 2rem;
-    font-size: 1.125rem;
-    line-height: 32px;      
-    justify-content: flex-end;
-    padding: 0 0.75rem;    
-    display: flex;
-    -ms-flex-preferred-size: 100%;
-    flex-basis: 100%;
-    
-  }
-
-  .blog-cards .pagination .pagination-btn {
-    display: flex;
-  }
-
-  .blog-cards .pagination > li {
-    display: flex;
-  }
-
-  .blog-cards .pagination >li .prev {
-    position: relative;
-    width: 7px;
-    height: 32px;
-    margin-right: 1rem;
-}
-
-.blog-cards .pagination >li .prev::before {
-  position: absolute;
-  left: 0;
-  top: 0;
-  content: "";    
-  width: 100%;
-  height: 100%;
-  background: url('../../icons/previous.svg');
-  background-size: contain;
-  background-position: 50%;
-  background-repeat: no-repeat;
-}
-
-  .blog-cards .pagination >li .next{
-    position: relative;
-    width: 7px;
-    height: 32px;
-    margin-left: 1rem;
-  }
-
-  .blog-cards .pagination >li .next::after {
-  position: absolute;
-  right: 0;
-  top: 0;
-  content: "";    
-  width: 100%;
-  height: 100%;
-  background: url('../../icons/next.svg');
-  background-size: contain;
-  background-position: 50%;
-  background-repeat: no-repeat;
-}
-}
-
-@media (min-width: 900px) {
-   .blog-cards > ul {      
-      display: grid;
-      grid-template-columns: repeat(3, 31.33%);
-      grid-gap: 24px;   
-      margin-bottom: 1rem;    
-      min-height: 0;         
-    }
-
-    
-.acsb-sr-only {
-  display: block !important;
-  position: absolute !important;
-  left: 0 !important;
-  top: 0 !important;
-  margin: -1px 0 0 -1px !important;
-  padding: 0 !important;
-  width: 1px !important;
-  height: 1px !important;
-  font-weight: normal !important;
-  font-size: 15px !important;
-  line-height: 1px !important;
-  overflow: hidden !important;
-  clip: rect(0, 0, 0, 0) !important;
-  border: 0 !important;
-  outline: 0 !important;
-  cursor: text !important;
-  text-transform: capitalize !important;
-  color: #000 !important;
-  background: #fff !important;
-}
-  }
-
-  @media (min-width: 1200px) {
-    .blog-cards > ul {
-      grid-template-columns: repeat(3, 31.20%);
-      grid-gap: 16px;   
-      
-    }    
-  }
