+++ blocks/cards/cards.css
@@ -29,3 +29,120 @@
   aspect-ratio: 4 / 3;
   object-fit: cover;
 }
+
+main {
+  --offst-left: 20px;
+}
+
+
+main a {
+    color: var(--color-gray-dark);    
+}
+
+@media (min-width: 400px) {
+.cards.article > ul {
+    list-style: none;
+    margin: 0;                    
+    font-size: 1em;
+    font-weight: 400;
+    line-height: 1.5;
+    display: block;
+    padding-inline-start: 0;
+    min-height: 100vh;
+  }
+  
+  .cards.article > ul > li {     
+    border: 1px solid var(--color-light-blue);
+    border-radius: 6px;
+    padding: 0;
+    margin-bottom: 1.5rem;    
+  }
+  
+  .cards.article .blogs-card-body {
+    padding: 1.5rem;
+    display: flex;    
+    flex-direction: column;    
+    justify-content: space-between;    
+    background-color: transparent;
+    min-height: 220px;  
+  } 
+  
+  .cards.article .blogs-card-image {
+    line-height: 0;
+    position: relative;
+    display: block;
+   
+    
+  }
+   
+  .cards.article > ul > li img {
+    border-top-left-radius: 6px;
+    border-top-right-radius: 6px;
+    aspect-ratio: 4 / 3;
+    max-width: 100%;
+    object-fit: cover
+  }
+  
+  .cards.article .blogs-card-body > h3{
+    font-size: var(--body-font-size-l);
+    margin-bottom: 10px;    
+    font-family: var(--heading-font-family);
+    font-weight: 800;
+    color: var(--text-color);
+  }
+
+  .cards.article .blogs-card-body >span{
+    font-size: var(--body-font-size-s);
+  }
+
+  .cards.article .blogs-card-category{
+    font-family: var(--heading-font-family);    
+    position: absolute;
+    line-height: 1.7;
+    top: 10px;
+    left: 20px;
+    border-radius: 14px;
+    display: inline-block;
+    padding: 0 14px;    
+    color: var(--text-color-inverted);    
+    max-width: calc(100% - 2 * var(--offset-left));
+  }
+
+  .cards.article .pagination{
+    margin-top: 2rem;
+    font-size: 1.125rem;
+    line-height: 32px;      
+    justify-content: flex-end;
+    padding: 0 0.75rem;    
+    display: flex;    
+    flex-basis: 100%;
+    
+  }
+
+  .cards.article .pagination .pagination-btn {
+    display: flex;
+  }
+
+  .cards.article .pagination > li {
+    display: flex;
+  }
+
+  .cards.article .pagination >li.prev, li.next {
+    position: relative;
+    width: 7px;
+    height: 32px;
+    margin-right: 1rem;
+}
+}
+
+
+@media (min-width: 900px) {
+  .cards.article > ul {      
+      display: grid;
+      grid-template-columns: repeat(3, 1fr);
+      grid-gap: 24px;   
+      margin-bottom: 1rem;    
+      min-height: 0;         
+    } 
+
+  }
